% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-dists.R
\name{allDistributions}
\alias{allDistributions}
\title{function to make list of all distributions and relevant info}
\usage{
allDistributions()
}
\value{
list of distributions - each distribution has elements 'distribution', 'name', 'bounds', and 'isInt'
}
\description{
function to make list of all distributions and relevant info
}

