% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-and-helpers.R
\name{Alteryx_distributions}
\alias{Alteryx_distributions}
\title{Get distributions supported by Alteryx tools}
\usage{
Alteryx_distributions()
}
\value{
vector of supported distributions
}
\description{
Get distributions supported by Alteryx tools
}
\examples{
   Alteryx_distributions()
}

