% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-and-helpers.R
\name{Alteryx_distributions_continuous}
\alias{Alteryx_distributions_continuous}
\title{Get continuous distributions supported by Alteryx tools}
\usage{
Alteryx_distributions_continuous()
}
\value{
vector of supported continuous distributions
}
\description{
Get continuous distributions supported by Alteryx tools
}
\examples{
   Alteryx_distributions_continuous()
}

