% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-and-helpers.R
\name{Alteryx_distributions_discrete}
\alias{Alteryx_distributions_discrete}
\title{Get discrete distributions supported by Alteryx tools}
\usage{
Alteryx_distributions_discrete()
}
\value{
vector of supported discrete distributions
}
\description{
Get discrete distributions supported by Alteryx tools
}
\examples{
   Alteryx_distributions_continuous()
}

