% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{bin_to_data}
\alias{bin_to_data}
\title{Converts binned data into vector of data}
\usage{
bin_to_data(bins, count = NULL, discrete = FALSE)
}
\arguments{
\item{bins}{dataframe with columns for id and count}

\item{count}{number of samples to draw}

\item{discrete}{whether the data is discrete (default is False)}
}
\value{
convert bins into vector of data
}
\description{
Converts binned data into vector of data
}

