% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{checkForMissingXVars}
\alias{checkForMissingXVars}
\title{Check for missing predictors in score data as compared to the model}
\usage{
checkForMissingXVars(model, data)
}
\arguments{
\item{model}{model object}

\item{data}{data to score}
}
\description{
Check for missing predictors in score data as compared to the model
}

