% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-chisq.R
\name{chi_sq}
\alias{chi_sq}
\title{Convert a fit to corresponding chisq value}
\usage{
chi_sq(fit_obj)
}
\arguments{
\item{fit_obj}{object fit by}
}
\value{
chi squared statistic
}
\description{
Convert a fit to corresponding chisq value
}
\examples{
\dontrun{
  chi_sq(fit_best_single(rnorm(100), "normal"))
  chi_sq(fit_best_single(rnorm(100), "triangle"))
}
}

