% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-and-helpers.R
\name{convert_dist}
\alias{convert_dist}
\title{Convert distribution name to R distribution function suffix}
\usage{
convert_dist(distribution)
}
\arguments{
\item{distribution}{string (or vector of strings) - name of distribution}
}
\value{
string (or vector of strings) - R-recognized distribution, has corresponding d|p|q|r functions
}
\description{
Convert distribution name to R distribution function suffix
}
\examples{
   convert_dist("normal")
   convert_dist(c("normal", "gamma", "lognormal"))
}

