% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFormula.R
\name{convertDist}
\alias{convertDist}
\title{Function converting distribution name to R-accepted distribution postfix
If non-supported distribution name entered, returns inputted name}
\usage{
convertDist(distName)
}
\arguments{
\item{distName}{name of distribiution}
}
\value{
R-supported distribution postfix string
}
\description{
Function converting distribution name to R-accepted distribution postfix
If non-supported distribution name entered, returns inputted name
}

