% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{convertTo01.character}
\alias{convertTo01.character}
\title{convert character of Yes/No vector to 0/1}
\usage{
\method{convertTo01}{character}(vec)
}
\arguments{
\item{vec}{vector to convert}
}
\value{
vector of 0/1
}
\description{
convert character of Yes/No vector to 0/1
}

