% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{convertTo01.default}
\alias{convertTo01.default}
\title{Convert default vector to 0/1}
\usage{
\method{convertTo01}{default}(vec)
}
\arguments{
\item{vec}{vector to convert}
}
\value{
vector of 0/1
}
\description{
Convert default vector to 0/1
}

