% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{convertTo01.factor}
\alias{convertTo01.factor}
\title{convert character vector to 0/1}
\usage{
\method{convertTo01}{factor}(vec)
}
\arguments{
\item{vec}{vector to convert}
}
\value{
vector of 0/1
}
\description{
convert character vector to 0/1
}

