% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{convertTo01}
\alias{convertTo01}
\title{Generic method for converting vector to 0/1}
\usage{
convertTo01(vec)
}
\arguments{
\item{vec}{vector to convert}
}
\value{
vector of 0/1
}
\description{
Generic method for converting vector to 0/1
}

