% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-chisq.R
\name{custom_chisq}
\alias{custom_chisq}
\title{Get chi squared value from MLE for custom mle function}
\usage{
custom_chisq(custom_mle)
}
\arguments{
\item{custom_mle}{named list containing 'data' vector, 'distribution' name,
and 'param_list' list}
}
\value{
chisq statistic with number of bins 1/4*length(data)^{3/5} -
  corresponding to average bin size used in fitdistrplus package
}
\description{
Get chi squared value from MLE for custom mle function
}
\examples{
\dontrun{
   custom_chisq(fit_best_single(rnorm(100), "triangular"))
}
}

