% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-chisq.R
\name{custom_chisq_discrete}
\alias{custom_chisq_discrete}
\title{Get chisq value from MLE for custom chisq function for discrete dists}
\usage{
custom_chisq_discrete(custom_mle)
}
\arguments{
\item{custom_mle}{named list containing 'data' vector, 'distribution' name,
and 'param_list' list}
}
\value{
chisq statistic with number of bins 1/4*length(data)^{3/5} -
  corresponding to average bin size used in fitdistrplus package
}
\description{
Get chisq value from MLE for custom chisq function for discrete dists
}

