% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling-processes.R
\name{data_process}
\alias{data_process}
\title{Code for processing data inputs}
\usage{
data_process(method, chunkSize, count, process, possible, type, id, value, name,
  roulette, data)
}
\arguments{
\item{method}{sampling method used: "MC" for Monte Carlo or "LH" for Latin Hypercube}

\item{chunkSize}{int > 0 - maximal number of records to be processed at time}

\item{count}{count}

\item{process}{sampling method for data - "entire" to sample records, "each" to sample from columns independently, "best" to fit then sample}

\item{possible}{vector of distribution options to be fit to data}

\item{type}{whether data is "binned", comes from "roulette" widget, or is standard}

\item{id}{name of id field for binned data}

\item{value}{name of value field for binned data}

\item{name}{name of output vector for roulette or binned data or parametric input}

\item{roulette}{json string for roulette data}

\item{data}{incomming data if process = "data"}
}
\value{
function writes to Alteryx output 1
}
\description{
Code for processing data inputs
}

