% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFormula.R
\name{distApply}
\alias{distApply}
\title{Apply desired function of distribution (given by distribution and prefixLetter) to all values
Takes parameters in df or constant params to apply to all values}
\usage{
distApply(distribution, type, df, valName, constParams, param1Name, param2Name,
  param3Name, param1, param2, param3)
}
\arguments{
\item{distribution}{distribution name}

\item{type}{distribution type to sample from. "pdf", "cdf", "icdf" (for inverse), "rand" (for random)}

\item{df}{incoming data}

\item{valName}{name of value field in data (not necessary for type = "rand")}

\item{param1Name}{name of field for param1 - null when constParam}

\item{param2Name}{name of field for param2 - null when constParam or distribution has 1 parameter}

\item{param3Name}{name of field for param3 - null when constParam or distribution has 1-2 parameters}

\item{param1}{value for param1 - null when !constParam}

\item{param2}{value for param2 - null when !constParam or distribution has 1 parameter}

\item{param3}{value for param3 - null when !constParam or distribution has 1-2 parameters}

\item{constParam}{bool giving whether params are constant across all records}
}
\value{
dataframe with 1 column "results" containing results from thet appropriate function applied to every value
}
\description{
Apply desired function of distribution (given by distribution and prefixLetter) to all values
Takes parameters in df or constant params to apply to all values
}

