% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-and-helpers.R
\name{doInChunks}
\alias{doInChunks}
\title{Apply function to each chunk if in Alteryx. If outside, run on full size}
\usage{
doInChunks(nOutput, total_size, chunk_size, names = NULL)
}
\arguments{
\item{nOutput}{int >= 0; if 0, run outside of Alteryx; if = 0, denotes output number in Alteryx}

\item{total_size}{total size of data}

\item{chunk_size}{maximal size of chunk}

\item{names}{names}
}
\value{
function
}
\description{
Apply function to each chunk if in Alteryx. If outside, run on full size
}

