% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling-processes.R
\name{each_process}
\alias{each_process}
\title{Code for processing when process = "each"}
\usage{
each_process(chunkSize, count, data)
}
\arguments{
\item{chunkSize}{int > 0 - maximal number of records to be processed at time}

\item{count}{number of samples to draw}

\item{data}{incomming data if process = "data"}
}
\value{
function writes to Alteryx output 1
}
\description{
Code for processing when process = "each"
}

