% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-and-helpers.R
\name{errorCheckBounds.binomial}
\alias{errorCheckBounds.binomial}
\title{Error check for binomial out of bounds}
\usage{
\method{errorCheckBounds}{binomial}(boundsvec)
}
\description{
Error check for binomial out of bounds
}

