% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-and-helpers.R
\name{errorCheckBounds.default}
\alias{errorCheckBounds.default}
\title{Default error check for bounds method}
\usage{
\method{errorCheckBounds}{default}(boundsVec)
}
\arguments{
\item{boundsVec}{vector of bounds}
}
\description{
Default error check for bounds method
}

