% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-and-helpers.R
\name{errorCheckBounds.gamma}
\alias{errorCheckBounds.gamma}
\title{Error check for gamma out of bounds}
\usage{
\method{errorCheckBounds}{gamma}(boundsVec)
}
\arguments{
\item{boundsVec}{vector of bounds}
}
\description{
Error check for gamma out of bounds
}

