% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-and-helpers.R
\name{errorCheckBounds.geometric}
\alias{errorCheckBounds.geometric}
\title{Error check for geometric out of bounds}
\usage{
\method{errorCheckBounds}{geometric}(boundsvec)
}
\description{
Error check for geometric out of bounds
}

