% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-and-helpers.R
\name{errorCheckBounds.lognormal}
\alias{errorCheckBounds.lognormal}
\title{Error check for lognormal out of bounds}
\usage{
\method{errorCheckBounds}{lognormal}(boundsVec)
}
\arguments{
\item{boundsVec}{vector of bounds}
}
\description{
Error check for lognormal out of bounds
}

