% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-and-helpers.R
\name{errorCheckBounds.pareto}
\alias{errorCheckBounds.pareto}
\title{Error check for pareto out of bounds}
\usage{
\method{errorCheckBounds}{pareto}(boundsVec)
}
\arguments{
\item{boundsVec}{vector of bounds}
}
\description{
Error check for pareto out of bounds
}

