% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-and-helpers.R
\name{errorCheckBounds.poisson}
\alias{errorCheckBounds.poisson}
\title{Error check for poisson out of bounds}
\usage{
\method{errorCheckBounds}{poisson}(boundsVec)
}
\arguments{
\item{boundsVec}{vector of bounds}
}
\description{
Error check for poisson out of bounds
}

