% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-and-helpers.R
\name{errorCheckBounds}
\alias{errorCheckBounds}
\title{Generic error check for illegal bounds; throws error if set bound is outside of absolute bound}
\usage{
errorCheckBounds(boundsVec)
}
\arguments{
\item{boundsVec}{vector of bounds}
}
\description{
Generic error check for illegal bounds; throws error if set bound is outside of absolute bound
}

