% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-and-helpers.R
\name{errorCheckParams.default}
\alias{errorCheckParams.default}
\title{Default error check; returns no error}
\usage{
\method{errorCheckParams}{default}(paramList)
}
\arguments{
\item{paramList}{list of parameters}
}
\description{
Default error check; returns no error
}

