% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-and-helpers.R
\name{errorCheckParams}
\alias{errorCheckParams}
\title{Generic check for illegal parameters; throws error if params out of bounds}
\usage{
errorCheckParams(paramList)
}
\arguments{
\item{paramList}{list of parameters}
}
\description{
Generic check for illegal parameters; throws error if params out of bounds
}

