% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-and-helpers.R
\name{errorCheckParams.triangular}
\alias{errorCheckParams.triangular}
\title{triangular distribution error check}
\usage{
\method{errorCheckParams}{triangular}(paramList)
}
\arguments{
\item{paramList}{list of parameters}
}
\description{
triangular distribution error check
}

