% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-and-helpers.R
\name{errorCheckParams.uniform}
\alias{errorCheckParams.uniform}
\title{uniform distribution error check}
\usage{
\method{errorCheckParams}{uniform}(paramList)
}
\arguments{
\item{paramList}{list of parameters}
}
\description{
uniform distribution error check
}

