% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-best-fits.R
\name{fapply}
\alias{fapply}
\title{Apply list of functions to single argument}
\usage{
fapply(fxns, arg)
}
\arguments{
\item{fxns}{list of functions}

\item{arg}{arg to repeatedly apply fxns to}
}
\value{
list of results from functions applied to args
}
\description{
Apply list of functions to single argument
}
\examples{
   fapply(c(min, max, mean), 1:5)
}

