% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-fitting.R
\name{fit_best_single}
\alias{fit_best_single}
\title{For specified distribution, find parameters for best-fitting distribution}
\usage{
fit_best_single(data, distribution)
}
\arguments{
\item{data}{data to fit distribution to}

\item{distribution}{name of the distribution to fit}
}
\value{
parameter list of MLE for distribution on data
}
\description{
For specified distribution, find parameters for best-fitting distribution
}
\examples{
   fit_best_single(rnorm(100), "normal")
   fit_best_single(rtriangle(100), "triangle")
}

