% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-fitting.R
\name{fit_dists.data.frame}
\alias{fit_dists.data.frame}
\title{Fit each column to best fitting distribution and distribution params}
\usage{
\method{fit_dists}{data.frame}(data, dist_list)
}
\arguments{
\item{data}{vector to fit}

\item{dist_list}{data to fit to}
}
\value{
list of distributions and best-fitting params
}
\description{
Fit each column to best fitting distribution and distribution params
}

