% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-fitting.R
\name{fit_dists.default}
\alias{fit_dists.default}
\title{Fit single vector across all distributions to best parameters}
\usage{
\method{fit_dists}{default}(data, dist_list)
}
\arguments{
\item{data}{vector to fit}

\item{dist_list}{data to fit to}
}
\value{
distributions and best-fitting params
}
\description{
Fit single vector across all distributions to best parameters
}

