% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-fitting.R
\name{fit_dists}
\alias{fit_dists}
\title{Generic method for fitting over vector of distributions}
\usage{
fit_dists(data, dist_list)
}
\arguments{
\item{data}{vector to fit}

\item{dist_list}{data to fit to}
}
\value{
best fits for distribution
}
\description{
Generic method for fitting over vector of distributions
}

