% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-fitting.R
\name{fit_dists2}
\alias{fit_dists2}
\title{Generic non-S3 method for distribution fitting}
\usage{
fit_dists2(data, dist_list)
}
\arguments{
\item{data}{vector to fit}

\item{dist_list}{data to fit to}
}
\description{
Generic non-S3 method for distribution fitting
}

