% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-fitting.R
\name{fit_info}
\alias{fit_info}
\title{Get info on fitting and gof}
\usage{
fit_info(data, distribution)
}
\arguments{
\item{data}{vector to fit to}

\item{distribution}{distribution to fit to}
}
\value{
list with elements 'distribution', 'params', and 'chisq'
}
\description{
Get info on fitting and gof
}

