% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-fitting.R
\name{fitdist_Alteryx.binomial}
\alias{fitdist_Alteryx.binomial}
\title{Fit binomial}
\usage{
\method{fitdist_Alteryx}{binomial}(x, ...)
}
\arguments{
\item{x}{object to fit}

\item{...}{additional parameters to pass.}
}
\value{
parameter list with elements:
   size
   prob
}
\description{
Fit binomial
}

