% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-fitting.R
\name{fitdist_Alteryx.default}
\alias{fitdist_Alteryx.default}
\title{Default method for fitting distribution
Uses fitdistrplus::fitdist without starting values to optimize}
\usage{
\method{fitdist_Alteryx}{default}(x, ...)
}
\arguments{
\item{x}{object to fit}

\item{...}{additional parameters to pass.}
}
\description{
Default method for fitting distribution
Uses fitdistrplus::fitdist without starting values to optimize
}

