% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-fitting.R
\name{fitdist_Alteryx.geom}
\alias{fitdist_Alteryx.geom}
\title{Fit geometric distribution}
\usage{
\method{fitdist_Alteryx}{geom}(x, ...)
}
\arguments{
\item{x}{object to fit}

\item{...}{additional parameters to pass.}
}
\value{
parameter list with elements:
   prob
}
\description{
Fit geometric distribution
}

