% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-fitting.R
\name{fitdist_Alteryx.pareto}
\alias{fitdist_Alteryx.pareto}
\title{Fit pareto distribution}
\usage{
\method{fitdist_Alteryx}{pareto}(x, ...)
}
\arguments{
\item{x}{object to fit}

\item{...}{additional parameters to pass.}
}
\value{
parameter list with elements:
   xm - min x val for distribution
   alpha - shape parameter
}
\description{
Fit pareto distribution
}

