% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-fitting.R
\name{fitdist_Alteryx.poisson}
\alias{fitdist_Alteryx.poisson}
\title{Fit poisson distribution}
\usage{
\method{fitdist_Alteryx}{poisson}(x, ...)
}
\arguments{
\item{x}{object to fit}

\item{...}{additional parameters to pass.}
}
\value{
parameter list with elements:
   lambda
}
\description{
Fit poisson distribution
}

