% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-fitting.R
\name{fitdist_Alteryx}
\alias{fitdist_Alteryx}
\title{Generic method for fitting distribution}
\usage{
fitdist_Alteryx(x, ...)
}
\arguments{
\item{x}{object to fit}

\item{...}{additional parameters to pass.}
}
\description{
Generic method for fitting distribution
}

