% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-fitting.R
\name{fitdist_Alteryx.triangle}
\alias{fitdist_Alteryx.triangle}
\title{Fit triangular distribution}
\usage{
\method{fitdist_Alteryx}{triangle}(x, ...)
}
\arguments{
\item{x}{object to fit}

\item{...}{additional parameters to pass.}
}
\value{
parameter list with elements:
   a - min x value
   b - man y value
   c - most likely x value
}
\description{
Fit triangular distribution
}

