% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-and-helpers.R
\name{get_chunk_sizes}
\alias{get_chunk_sizes}
\title{Give vector of chunk sizes}
\usage{
get_chunk_sizes(chunk_size, total_size)
}
\arguments{
\item{chunk_size}{maximal size of chunk}

\item{total_size}{sum of chunk sizes}
}
\value{
vector of chunk_sizes
}
\description{
Give vector of chunk sizes
}
\examples{
   get_chunk_sizes(2,5)
   get_chunk_sizes(5,2)
}

