% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{getErrorVec}
\alias{getErrorVec}
\title{Function to get error vector}
\usage{
getErrorVec(the.data, mod.obj)
}
\arguments{
\item{the.data}{data}

\item{mod.obj}{model object}
}
\value{
vector of errors
}
\description{
Function to get error vector
}

