% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-from-data.R
\name{getSampleSizes}
\alias{getSampleSizes}
\title{Get sample sizes for chunks}
\usage{
getSampleSizes(chunkSize, totalSize, sampleCount, replace = FALSE,
  strat = FALSE)
}
\arguments{
\item{chunkSize}{maximal chunksize}

\item{totalSize}{total data size}

\item{sampleCount}{number of total samples to draw}

\item{replace}{bool - sample with replacement}

\item{strat}{bool - stratified sampling}
}
\description{
Get sample sizes for chunks
}

