% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{getScore}
\alias{getScore}
\title{Helper function for simulateNonGLM for scores < 0 or > 1}
\usage{
getScore(x)
}
\arguments{
\item{x}{value to fix score of}
}
\value{
proper score
}
\description{
Helper function for simulateNonGLM for scores < 0 or > 1
}

