% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{getTarVarName}
\alias{getTarVarName}
\title{Extract target variable name from model object}
\usage{
getTarVarName(model)
}
\arguments{
\item{model}{model from which to extract target variable}
}
\value{
string - name of model's target variable
}
\description{
Extract target variable name from model object
}

