% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{loadPackageForScoring}
\alias{loadPackageForScoring}
\title{Load any needed packages based on the class of the model}
\usage{
loadPackageForScoring(model)
}
\arguments{
\item{model}{model to score}
}
\description{
Load any needed packages based on the class of the model
}

