% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapReduce.R
\name{mapReduce}
\alias{mapReduce}
\title{Map reduce code to help with processing of chunks}
\usage{
mapReduce(nameInput, chunkSize, totalSize, data)
}
\arguments{
\item{nameInput}{if data is coming from Alteryx, name of the input data is coming from}

\item{chunkSize}{maximal size of chunk}

\item{totalSize}{total size of data}

\item{data}{dataframe outside of Alterx, not dataframe if inside Alteryx}
}
\value{
function taking two function arguments - a map function and a reduce function and applying them as expected
}
\description{
Map reduce code to help with processing of chunks
}
\examples{
\dontrun{
   mapReduce(NULL, 15, 100,
     data.frame(x = 1:100)
   )(sum, function(x){sum(unlist(x))})
   mapReduce(NULL, 15, 100,
     data.frame(x = 1:100)
   )(function(x) {sum(x^2)}, function(x) {sum(unlist(x))})
}
}

