% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-from-data.R
\name{nonStratSampling}
\alias{nonStratSampling}
\title{Get sample sizes for simple (non-stratified) sampling}
\usage{
nonStratSampling(chunkSizes, totalSize, sampleCount, replace = FALSE)
}
\arguments{
\item{chunkSizes}{vector of chunk sizes}

\item{totalSize}{total data size}

\item{sampleCount}{number of total samples to draw}

\item{replace}{bool - sample with replacement}
}
\value{
vector of sample sizes
}
\description{
Get sample sizes for simple (non-stratified) sampling
}

