% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{prepareDataForScoring}
\alias{prepareDataForScoring}
\title{Score a model}
\usage{
prepareDataForScoring(model, data)
}
\arguments{
\item{model}{model object}

\item{data}{data to score}

\item{...}{other parameters}
}
\description{
Score a model
}

