% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-and-helpers.R
\name{qparetoAlteryx}
\alias{qparetoAlteryx}
\title{pareto inverse CDF function}
\usage{
qparetoAlteryx(p, shape, scale)
}
\arguments{
\item{p}{probability}

\item{shape}{shape parameter (minimum x)}

\item{scale}{scale parameter}
}
\value{
value
}
\description{
pareto inverse CDF function
}

