% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparam-sampling.R
\name{quantile_sample}
\alias{quantile_sample}
\title{Quantile function sampling}
\usage{
quantile_sample(distribution, quantile_vector, param_list)
}
\arguments{
\item{distribution}{distribution to sample from}

\item{quantile_vector}{vector of quantiles}

\item{param_list}{list or vector of parameters}
}
\value{
vector of sampled data points
}
\description{
Quantile function sampling
}
\examples{
   quantile_sample("normal", c(0,.25,.5,.75,1), list(mean = 0, sd = 1))
   quantile_sample("lognormal", seq(0,1,by=.1), list(meanlog = 0, sdlog = 1))
}

