% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-and-helpers.R
\name{rparetoAlteryx}
\alias{rparetoAlteryx}
\title{pareto random sampling function}
\usage{
rparetoAlteryx(n, shape, scale)
}
\arguments{
\item{n}{number of observations.}

\item{shape}{shape parameter (minimum x)}

\item{scale}{scale parameter}
}
\value{
vector of random values
}
\description{
pareto random sampling function
}

