% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity-analysis.R
\name{sa_importance}
\alias{sa_importance}
\title{Compute Importance Index}
\usage{
sa_importance(data, inputs, outputs)
}
\arguments{
\item{data}{dataset to plot from}

\item{inputs}{vector of strings - x axis variables}

\item{outputs}{vector of strings - y axis variables}
}
\value{
matrix of importance indices - one for each pair of input/output
}
\description{
Compute Importance Index
}

