% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity-analysis.R
\name{sa_pr}
\alias{sa_pr}
\title{Partial Correlation Coefficient}
\usage{
sa_pr(data, inputs, outputs)
}
\arguments{
\item{data}{dataset to plot from}

\item{inputs}{vector of strings - x axis variables}

\item{outputs}{vector of strings - y axis variables}
}
\value{
matrices of partial correlation coefficients
}
\description{
Partial Correlation Coefficient
}

