% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity-analysis.R
\name{sa_r}
\alias{sa_r}
\title{Compute Pearson's R matrix of correlation coefficients}
\usage{
sa_r(data, inputs, outputs)
}
\arguments{
\item{data}{dataset to plot from}

\item{inputs}{vector of strings - x axis variables}

\item{outputs}{vector of strings - y axis variables}
}
\description{
Compute Pearson's R matrix of correlation coefficients
}

