% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity-analysis.R
\name{sa_scat}
\alias{sa_scat}
\title{Scatter plotting}
\usage{
sa_scat(data, inputs, outputs)
}
\arguments{
\item{data}{dataset to plot from}

\item{inputs}{vector of strings - x axis variables}

\item{outputs}{vector of strings - y axis variables}
}
\value{
collection of scatterplots - one for each pair of input/output
}
\description{
Scatter plotting
}

