% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-best-fits.R
\name{sample_best}
\alias{sample_best}
\title{Find best fit for each column and sample from it}
\usage{
sample_best(data, dist_list = Alteryx_distributions_continuous(),
  type = "MC")
}
\arguments{
\item{data}{dataframe to fit columns of}

\item{dist_list}{list of distributions whose fits to consider}

\item{type}{"MC" or "LH" for Monte Carlo or Latin Hypercube sampling}
}
\value{
function that takes size argument and gives dataframe of samples
}
\description{
Find best fit for each column and sample from it
}

