% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-from-data.R
\name{sample_df}
\alias{sample_df}
\title{Sample entire records from dataset}
\usage{
sample_df(df, replace = TRUE, ...)
}
\arguments{
\item{df}{dataframe from which to sample}

\item{replace}{bool - whether to sample with replacement}

\item{...}{additional parameters for 'sample' function}
}
\value{
samples from dataframe
}
\description{
Sample entire records from dataset
}

