% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-from-data.R
\name{sample_df_indep}
\alias{sample_df_indep}
\title{Sample from dataset, sampling from columns independently}
\usage{
sample_df_indep(df, replace = TRUE, ...)
}
\arguments{
\item{df}{dataframe from which to sample}

\item{replace}{bool - whether to sample with replacement}

\item{...}{additional parameters for 'sample' function}
}
\value{
samples from dataframe
}
\description{
Sample from dataset, sampling from columns independently
}

