% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparam-sampling.R
\name{sample_hyper}
\alias{sample_hyper}
\title{Samples flexible parameters from distribution}
\usage{
sample_hyper(distribution = "uniform", quantile_count, param_list = list(min
  = 0, max = 1), inf.remove = TRUE)
}
\arguments{
\item{distribution}{name of distribution}

\item{quantile_count}{desired number of equidistant quantiles}

\item{param_list}{first parameter for 'distribution' - params for distribution}

\item{inf.remove}{if TRUE, removes +Inf and -Inf value from resulting vector}
}
\value{
Vector of length quantile_count, quantile_count - 1, or quantile_count - 2,
   depending on inf.remove and number of inf values
}
\description{
Samples flexible parameters from distribution
}
\examples{
 sample_hyper(distribution = "normal", quantile_count = 11, param_list =
 list(mean = 0, sd = 1))
}

