% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparam-sampling.R
\name{sample_hyper_join}
\alias{sample_hyper_join}
\title{Joins hyperparameter to existing data}
\usage{
sample_hyper_join(data, name, ...)
}
\arguments{
\item{data}{data to join to hyperparam}

\item{name}{name of flexible paramter to cartesian product with}

\item{...}{additional arguments to pass to \code{sample_hyper}}
}
\value{
Vector of length quantile_count, quantile_count - 1, or quantile_count - 2,
   depending on inf.remove and number of inf values
}
\description{
Joins hyperparameter to existing data
}

