% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric-sampling.R
\name{sample_LH_from_dist}
\alias{sample_LH_from_dist}
\title{Latin Hypercube Parametric Sampling}
\usage{
sample_LH_from_dist(distribution, param_list)
}
\arguments{
\item{distribution}{- name of distribution from which to sample}

\item{param_list}{list or vector of parameters for distribution}
}
\value{
vector of length n samples from distribution
}
\description{
Latin Hypercube Parametric Sampling
}
\examples{
  sample_LH_from_dist("normal", c(0, 1))
  sample_LH_from_dist("normal", list(mean = 0, sd = 1))
}

