% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric-sampling.R
\name{sample_MC_from_dist}
\alias{sample_MC_from_dist}
\title{Monte Carlo Parametric Sampling}
\usage{
sample_MC_from_dist(distribution, param_list)
}
\arguments{
\item{distribution}{- name of distribution from which to sample}

\item{param_list}{list or vector of parameters for distribution}
}
\value{
function taking a size argument to return number of samples
}
\description{
Monte Carlo Parametric Sampling
}
\examples{
  sample_MC_from_dist("normal", list(0, 1))
  sample_MC_from_dist("normal", list(mean = 0, sd = 1))
  sample_MC_from_dist("hyper", list(m = 10, n = 7, k = 8))
  sample_MC_from_dist("hyper", list(10, 7, 8))(10)
}

