% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity-analysis.R
\name{saPr2}
\alias{saPr2}
\title{Partial Correlation Coefficients}
\usage{
saPr2(data)
}
\arguments{
\item{data}{dataframe of data}

\item{inputs}{vector of names for independent variable}

\item{outputs}{vector of dependent variable names}
}
\value{
list of matrices of partial correlations
}
\description{
Partial Correlation Coefficients
}

