% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity-analysis.R
\name{saPrSingle}
\alias{saPrSingle}
\title{Partial Correlation Coefficients}
\usage{
saPrSingle(data, inputs, output)
}
\arguments{
\item{data}{dataframe of data}

\item{inputs}{vector of names for independent variable}

\item{output}{single dependent variable name}
}
\value{
matrix of partial correlations
}
\description{
Partial Correlation Coefficients
}

