% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity-analysis.R
\name{saR}
\alias{saR}
\title{Pearson's R}
\usage{
saR(data, inputs, outputs)
}
\arguments{
\item{data}{dataframe to evaluate importances from}

\item{inputs}{vector of independent variable names}

\item{outputs}{vector of dependent variable names}
}
\value{
matrix of correlation coefficients
}
\description{
Pearson's R
}

