% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity-analysis.R
\name{saScat}
\alias{saScat}
\title{Function to make collection of scatter plots}
\usage{
saScat(data, inputs, outputs)
}
\arguments{
\item{data}{data for plot}

\item{inputs}{vector of independent variables}

\item{outputs}{vector of dependent variables}
}
\value{
grid of scatterplots
}
\description{
Function to make collection of scatter plots
}

