% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{scoreProcess}
\alias{scoreProcess}
\title{Code to execute core of simulation scoring code}
\usage{
scoreProcess(model, isGLM, mult, totalCountSim, validation, seed, chunkSize,
  scoreName)
}
\arguments{
\item{model}{model object to score from}

\item{isGLM}{boolean predicting whether object will be glm/lm class}

\item{mult}{number of draws from error distribution for each record}

\item{totalCountSim}{total number of records to be scored}

\item{validation}{validation data incoming}

\item{seed}{random seed}

\item{chunkSize}{maximal chunk size for scored data}

\item{scoreName}{name to give to scored variable}
}
\value{
nothing - function writes out score df to Alteryx
}
\description{
Code to execute core of simulation scoring code
}

