% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{simGLM}
\alias{simGLM}
\title{Repurpose {stats} package 'simulate' function to work on new data}
\usage{
simGLM(mod.obj, nsim)
}
\arguments{
\item{nsim}{number of draws per record}

\item{object}{model object to score from}
}
\value{
function taking new data and simulating scores
}
\description{
Repurpose {stats} package 'simulate' function to work on new data
}

