% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{simNonGLM}
\alias{simNonGLM}
\title{function to draw from error distribution and add to predictions (assumes homoscedasticity)}
\usage{
simNonGLM(mod.obj, errors, nsim)
}
\arguments{
\item{mod.obj}{model object}

\item{errors}{error vector to draw from}

\item{nsim}{number of draws per record}
}
\value{
function taking in new data and giving results
}
\description{
function to draw from error distribution and add to predictions (assumes homoscedasticity)
}

