% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{simulateGLM}
\alias{simulateGLM}
\title{Repurpose {stats} package 'simulate' function to work on new data}
\usage{
simulateGLM(object, nsim)
}
\arguments{
\item{object}{model object to score from}

\item{nsim}{number of draws per record}
}
\value{
function taking new data and simulating scores
}
\description{
Repurpose {stats} package 'simulate' function to work on new data
}

