% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{simulateNonGLM}
\alias{simulateNonGLM}
\title{Simulation function assuming homoscedasticity}
\usage{
simulateNonGLM(object, nsim, errorVec)
}
\arguments{
\item{object}{model object to score from}

\item{nsim}{number of draws per record}

\item{errorVec}{error vectorf}
}
\value{
function taking new data and simulating scores
}
\description{
Simulation function assuming homoscedasticity
}

