% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity-analysis.R
\name{singleScat}
\alias{singleScat}
\title{Function to make single scatter plot}
\usage{
singleScat(data, input, output)
}
\arguments{
\item{data}{data for plot}

\item{input}{variable name}

\item{output}{variable name}
}
\value{
scatter plot
}
\description{
Function to make single scatter plot
}

