% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-from-data.R
\name{stratSampling}
\alias{stratSampling}
\title{Get sample sizes for stratified sampling}
\usage{
stratSampling(chunkSize, totalSize, sampleCount)
}
\arguments{
\item{chunkSize}{maximal chunk size}

\item{totalSize}{total data size}

\item{sampleCount}{number of total samples to draw}
}
\value{
vector of sample sizes
}
\description{
Get sample sizes for stratified sampling
}

