% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{string_to_bin}
\alias{string_to_bin}
\title{Converts json string to dataframe with columns 'value' and 'count'}
\usage{
string_to_bin(incoming, bin_split = ",", id_split = ":")
}
\arguments{
\item{incoming}{json string with ids and values}

\item{bin_split}{splitting character between bins}

\item{id_split}{splitting character between id and bins}
}
\value{
dataframe of ids and values
}
\description{
Converts json string to dataframe with columns 'value' and 'count'
}
\examples{
   string_to_bin('{"0": 1, "1": 10, "2": 30}')
}

