% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling-processes.R
\name{tool_process}
\alias{tool_process}
\title{Code for sampling tool in Alteryx}
\usage{
tool_process(method, chunkSize, seed, count, distribution, params, bounds,
  process, possible, type, id, value, name, roulette, dataName, sampleSource,
  replace, totalSize)
}
\arguments{
\item{method}{sampling method used: "MC" for Monte Carlo or "LH" for Latin Hypercube}

\item{chunkSize}{int > 0 - maximal number of records to be processed at time}

\item{seed}{seed to intialize tool with}

\item{count}{number of samples to draw}

\item{distribution}{valid distribution name}

\item{params}{parameter list for distribution}

\item{bounds}{vector - first element is lower bound, second is upper. Closed intervals only}

\item{process}{sampling method for data - "entire" to sample records, "each"
to sample from columns independently, "best" to fit then sample}

\item{possible}{vector of distribution options to be fit to data}

\item{type}{whether data is "binned", comes from "roulette" widget, or is standard}

\item{id}{name of id field for binned data}

\item{value}{name of value field for binned data}

\item{name}{name of output vector for roulette or binned data or parametric input}

\item{roulette}{json string for roulette data}

\item{dataName}{incomming data if process = "data"}

\item{sampleSource}{sample source}

\item{replace}{whether or not to sample with replacement}

\item{totalSize}{total size}
}
\value{
function writes to Alteryx output 1
}
\description{
Code for sampling tool in Alteryx
}

