% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-fitting.R
\name{try_fit_best_single}
\alias{try_fit_best_single}
\title{Apply best fit function and catch potential errors
Necessary for impossible function fitting
Example: fitting lognormal on negative values}
\usage{
try_fit_best_single(data, distribution)
}
\arguments{
\item{data}{vector to fit to}

\item{distribution}{distribution to fit to}
}
\value{
fit object or NA (if can't be fit)
}
\description{
Apply best fit function and catch potential errors
Necessary for impossible function fitting
Example: fitting lognormal on negative values
}

