% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling-processes.R
\name{best_process}
\alias{best_process}
\title{Code for processing when process = "best"}
\usage{
best_process(method, chunkSize, count, data, dataName, possible)
}
\arguments{
\item{method}{sampling method used: "MC" for Monte Carlo or "LH" for Latin Hypercube}

\item{chunkSize}{int > 0 - maximal number of records to be processed at time}

\item{count}{number of samples to draw}

\item{data}{incomming data if process = "data"}

\item{dataName}{name of data input incoming}

\item{possible}{vector of distribution options to be fit to data}

\item{totalSize}{total size}
}
\value{
function writes to Alteryx output 1
}
\description{
Code for processing when process = "best"
}

