% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{removeMissingLevels}
\alias{removeMissingLevels}
\title{Remove missing levels from score data based on model data}
\usage{
removeMissingLevels(model, data)
}
\details{
@param model model object
 @param data data to score
}

