% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{scoreErrorCheck}
\alias{scoreErrorCheck}
\title{Error check to ensure model class matches "isGLM"}
\usage{
scoreErrorCheck(model, isLM)
}
\arguments{
\item{model}{model to check}

\item{isGLM}{boolean to compare}
}
\value{
no return - just throws error if mismatch
}
\description{
Error check to ensure model class matches "isGLM"
}

