% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{scoreProcess}
\alias{scoreProcess}
\title{Function to score model with added samples from an error distribution}
\usage{
scoreProcess(model, isGLM, mult, totalCountSim, validationData, seed, chunkSize)
}
\arguments{
\item{model}{model object to score against}

\item{isGLM}{boolean of whether model is of class lm/glm}

\item{mult}{number of samples from error distribution per record}

\item{totalCountSim}{total number of incoming records}

\item{validationData}{dataframe of validation data}

\item{seed}{random seed}

\item{chunkSize}{maximal size of incoming data chunk}
}
\value{
function taking data and scoring on it, sampling from error dist
}
\description{
Function to score model with added samples from an error distribution
}

